code re-factor
