+++ src/modules/contact.js
@@ -1,4 +1,6 @@
 /* eslint-disable linebreak-style */
+
+
 const contact = () => {
   const content = document.getElementById('content');
